Add Additional Docs to RevokedHandler Comments
We ran into some issues using the revoked handler; we were returning back the provided list of topic partition offsets, instead of an empty list.  It was difficult to find documentation on how this method worked, so we were hoping to clarify a bit more about how the return value works for the SetPartitionsRevokedHandler and how it interacts with partition assignment.