fix: remove breaking change to IConsumer
A method was added to the IConsumer interface by #2027 that resulted in
a breaking change in the 2.1 release. This replaces the interface
method and the implementation on Consumer with an extension method on
IConsumer to ensure that existing implementations of IConsumer remain
unbroken and have the same functionality as the default Consumer
implementation.